Fix review comments - Change table parsing - Create dedicated conftest for the test - Create file contains help function
